requirements.txt
selenium
streamlit
google-generativeai
requests
